requirements.txt
# Core dependencies for Research Paper Multi-Agent System
streamlit>=1.28.0
google-generativeai>=0.3.0
openai>=1.0.0
arxiv>=2.0.0
PyPDF2>=3.0.1
requests>=2.31.0
python-dotenv>=1.0.0

# Audio generation dependencies (Google TTS)
google-cloud-texttospeech>=2.16.0
gtts>=2.4.0
pydub>=0.25.1
